@@ -3,7 +3,7 @@
     "description": "Generate 64bit UUID with 45bit timestamp, 4bit machine id and 14bit serial number.",
     "type": "library",
     "require": {
-        "php": ">=7.0.0"
+        "php": ">=7.4"
     },
     "license": "MIT",
     "authors": [
